// src/types.d.ts
export interface Item {
    TestMode: boolean;
    ProductID: number;
    ManufacturerID: number;
    ItemID: string;
    ItemName: string;
    Description: string;
    Notes: string;
    Dimensions: string;
    BasePrice: number;
    SpecialPrice: number;
    OrderMinimumQuantity: number;
    ContainerMinQty: number;
    OrderMultipleQuantity: number;
    OnHandQuantity: number;
    Category?: null;
    IntroDate: string;
    BackOrderDate: string;
    UPC: string;
    PriceLevel1: number;
    PriceLevel2: number;
    PriceLevel3: number;
    PriceLevel4: number;
    PriceLevel5: number;
    PriceLevel6: number;
    PriceLevel7: number;
    PriceLevel8: number;
    PriceLevel9: number;
    PieceBox?: null;
    Cubes: number;
    UnitOfMeasure: string;
    UDF1: string;
    UDF2: string;
    UDF3: string;
    UDF4: string;
    UDF5: string;
    AdditionalImageCount?: number | null;
    PhotoName: string;
    AppendProductModifiers: boolean;
    Discontinued: boolean;
    CreatedOn: string;
    CreatedBy?: null;
    ModifiedOn: string;
    ModifiedBy: string;
    DeletedOn?: null;
    DeletedBy?: null;
    IsDeleted: boolean;
    PriceLevel10: number;
    PriceLevel11: number;
    PriceLevel12: number;
    PriceLevel13: number;
    PriceLevel14: number;
    PriceLevel15: number;
    PriceLevel16: number;
    PriceLevel17: number;
    PriceLevel18: number;
    PriceLevel19: number;
    PriceLevel20: number;
    Weight: number;
    DimensionsMetric?: null;
    Source: string;
    InventoryStatus?: null;
    CategoryIDs?: null;
    CatalogCode: string;
    CatalogName: string;
    udf6: string;
    udf7: string;
    udf8: string;
    udf9: string;
    udf10: string;
    udf11: string;
    udf12: string;
    udf13: string;
    udf14: string;
    udf15: string;
    udf16: string;
    udf17: string;
    udf18: string;
    udf19: string;
    udf20: string;
    reportCategory: string;
    additionalPhotos: string;
    showRelatedFor?: null;
    showChildFor?: null;
}